refactor: change mirinae hook file names
