+++ packages/mirinae/src/foundation/icons/config.ts
@@ -1,6 +1,6 @@
 export const ANIMATION_TYPE = {
     spin: 'spin',
     reserveSpin: 'reserve-spin',
-};
+} as const;
 
 export type AnimationType = typeof ANIMATION_TYPE[keyof typeof ANIMATION_TYPE];
