+++ packages/mirinae/src/hooks/use-context-menu-controller/use-context-menu-controller.ts
@@ -6,9 +6,9 @@ import {
 
 import { isEmpty } from 'lodash';
 
-import type { MenuAttachHandler } from '@/hooks/context-menu-controller/context-menu-attach';
-import { useContextMenuAttach } from '@/hooks/context-menu-controller/context-menu-attach';
-import { useContextMenuStyle } from '@/hooks/context-menu-style';
+import type { MenuAttachHandler } from '@/hooks/use-context-menu-controller/use-context-menu-attach';
+import { useContextMenuAttach } from '@/hooks/use-context-menu-controller/use-context-menu-attach';
+import { useContextMenuStyle } from '@/hooks/use-context-menu-style/use-context-menu-style';
 import type { MenuItem } from '@/inputs/context-menu/type';
 import { getTextHighlightRegex } from '@/utils/helpers';
