+++ packages/mirinae/src/inputs/dropdown/select-dropdown/PSelectDropdown.stories.ts
@@ -3,7 +3,7 @@ import { reactive, toRefs } from 'vue';
 import type { Meta, StoryObj } from '@storybook/vue';
 import type { ComponentProps } from 'vue-component-type-helpers';
 
-import { useProxyValue } from '@/hooks/proxy-state';
+import { useProxyValue } from '@/hooks/use-proxy-state/use-proxy-state';
 import PButton from '@/inputs/buttons/button/PButton.vue';
 import PToggleButton from '@/inputs/buttons/toggle-button/PToggleButton.vue';
 import { menuItems } from '@/inputs/context-menu/mock';
